Added back-propagation algorithm

+++ activation.go
@@ -5,5 +5,5 @@ import "math"
 // Activation functions for the neural networks
 
 func sigmoid(input float64) float64 {
-	return 1.0 / (1.0 + math.Exp(-input))
+	return float64(1.0) / (1.0 + math.Exp(-input))
 }
